+++ Block.h
@@ -5,7 +5,7 @@ class Block: public Entity{
 private:
 	int lifeLeft;
 public:
-	Block(float newX, float newY, int newWidth,
+	Block(int newX, int newY, int newWidth,
 		int newHeight, int newLifeLeft);
 	int getLifeLeft();
 	void decrementLifeLeft();
